Manage dataproc widget  (#4)
This PR adds cluster listing and filter by label to the %manage_dataproc widget. All widgets within the %manage_dataproc widget are now using  ipyveutify instead of ipywidgets. The storing and restoring endpoints and widgets logic has also changed. Previously, endpoints and sessions created with the %manage_dataproc widget could not be restored because the AddEndpointWidget and CreateSessionWidget classes in Sparkmagic controlled adding endpoints /sessions to sparkcontroller from the widget. This PR contains new AddEndpointWidget and CreateSessionWidget classes that are specific to DataprocMagic so now sessions and endpoints created with the widget are saved/restored.